stream update

+++ include/bidStrategy.cpp
@@ -3,7 +3,7 @@
 #include "advert.h"
 int bidStrategy(advert * current_ad)
 {
-  int price = rand() % 2;
+  double price = rand() % 2;
   current_ad->price = price;
   return 0;
 }
